refactor: Move trivy package under pkg/plugin
