+++ pkg/plugin/trivy/converter.go
