+++ pkg/plugin/trivy/model.go
