+++ pkg/plugin/trivy/model_test.go
