+++ pkg/plugin/trivy/plugin.go
